requirements.txt
flask==3.0.0
selenium==4.15.2
pillow==10.1.0
webdriver-manager==4.0.1
flask-caching==2.1.0
python-dotenv==1.0.0